MPC_Common: Signedness warnings

+++ mpcframework/MPC_Common/sctk_datastructures/utarray.h
@@ -261,7 +261,7 @@ typedef struct {
                         : NULL))
 #define utarray_back(a) (((a)->i) ? (_utarray_eltptr(a,(a)->i-1)) : NULL)
 #define utarray_eltidx(a, e)                                                   \
-  (((char *)(e) >= (a)->d) ? (((char *)(e) - (a)->d) / (a)->icd.sz) : -1)
+  (((char *)(e) >= (a)->d) ? (long long)(((char *)(e) - (a)->d) / (a)->icd.sz) : -1)
 
 /* last we pre-define a few icd for common utarrays of ints and strings */
 static void utarray_str_cpy(void *dst, const void *src) {
